requirements.txt
ffmpeg_python==0.2.0
pytube==15.0.0
Requests==2.31.0
youtube_transcript_api==0.6.1
